Use generic for `Printk` signature.
Previously the minimal example used `Printk1()` that specifically took an
integer as its parameter. This was done because dntc had a bug that did not
allow for a generic method in another assembly to be looked up correctly in
the transpiler.

The bug was fixed, and `Printk` now uses a generic argument to not have to
define different methods for different argument types.